+++ frontend/src/features/layout/components/DesktopNavLinks.tsx
@@ -30,7 +30,7 @@ export function DesktopNavLinks({ activePage }: DesktopNavLinksProps) {
   useEffect(() => {
     router.prefetch(paths.chat.getHref())
     router.prefetch(paths.code.getHref())
-    router.prefetch(paths.flow.getHref())
+    router.prefetch(paths.feed.getHref())
     if (isWikiEnabled) {
       router.prefetch(paths.wiki.getHref())
     }
@@ -61,7 +61,7 @@ export function DesktopNavLinks({ activePage }: DesktopNavLinksProps) {
         label: t('common:navigation.flow'),
         onClick: () => {
           startTransition(() => {
-            router.push(paths.flow.getHref())
+            router.push(paths.feed.getHref())
           })
         },
       },
