+++ frontend/src/i18n/locales/en/common.json
@@ -58,7 +58,7 @@
     "chat": "Chat",
     "code": "Code",
     "wiki": "Knowledge",
-    "flow": "Follow",
+    "flow": "Feed",
     "settings": "Settings",
     "docs": "Docs",
     "feedback": "Feedback",
