+++ frontend/src/i18n/locales/en/feed.json
@@ -1,11 +1,11 @@
 {
-  "title": "Follow",
-  "subtitle": "AI agent activities and execution feed",
+  "title": "Feed",
+  "subtitle": "AI-powered information tracking and updates",
   "tab_timeline": "Feed",
   "tab_config": "My Subscriptions",
   "my_flows": "My Subscriptions",
   "execution_timeline": "Execution Timeline",
-  "create_flow": "Create Flow",
+  "create_subscription": "Create Subscription",
   "create_first_flow": "Create your first Flow",
   "edit_flow": "Edit Flow",
   "edit_flow_desc": "Modify your flow configuration",
