+++ frontend/src/i18n/locales/zh-CN/feed.json
@@ -5,7 +5,7 @@
   "tab_config": "我的订阅",
   "my_flows": "我的订阅",
   "execution_timeline": "执行历史",
-  "create_flow": "新建订阅",
+  "create_subscription": "新建订阅",
   "create_first_flow": "创建第一个关注",
   "edit_flow": "编辑关注",
   "edit_flow_desc": "修改关注配置",
